requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
gunicorn==21.2.0 